+++ sdk/eventhub/azure-eventhub/azure/eventhub/_pyamqp/_connection.py
@@ -18,6 +18,7 @@ from .performatives import OpenFrame, CloseFrame
 from .constants import (
     PORT,
     SECURE_PORT,
+    WEBSOCKET_PORT,
     MAX_CHANNELS,
     MAX_FRAME_SIZE_BYTES,
     HEADER_FRAME,
@@ -99,6 +100,14 @@ class Connection(object):
             self._port = PORT
         self.state = None  # type: Optional[ConnectionState]
 
+        # Custom Endpoint
+        custom_endpoint_address = kwargs.get("custom_endpoint_address")
+        custom_endpoint = None
+        if custom_endpoint_address:
+            custom_parsed_url = urlparse(custom_endpoint_address)
+            custom_port = custom_parsed_url.port or WEBSOCKET_PORT
+            custom_endpoint = "{}:{}{}".format(custom_parsed_url.hostname, custom_port, custom_parsed_url.path)
+
         transport = kwargs.get('transport')
         self._transport_type = kwargs.pop('transport_type', TransportType.Amqp)
         if transport:
@@ -111,6 +120,7 @@ class Connection(object):
             self._transport = sasl_transport(
                 host=endpoint,
                 credential=kwargs['sasl_credential'],
+                custom_endpoint=custom_endpoint,
                 **kwargs
             )
         else: